Outsourced executeCommand() from check-diffs.ts
